arm64: dts: r8a7795: Add SYSC PM domains
Add a device node for the System Controller.
Hook up the Cortex-A57 CPU cores and the Cortex-A57 and Cortex A53 L2
caches/SCUs to their respective PM domains.
Hook up all devices that are part of the CPG/MSSR Clock Domain to the
SYSC "always-on" PM Domain.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
---
v4:
  - Use "<&sysc R8A7795_PD_ALWAYS_ON>" for recently added can0 and can1
    devices,

v3:
  - Drop power area hiearchy from DT,
  - Switch to "#power-domain-cells = <1>",
  - Use "<&sysc R8A7795_PD_ALWAYS_ON>" for devices in the always-on
    power area,
  - Drop fallback compatibility strings,

v2:
  - New.