v4l: vsp1: Reset the DRM pipeline when stopping the stream
When the stream is stopped all inputs need to be reset to NULL in order
to make them configurable again when restarting the stream. Don't use
vsp1_pipeline_reset() for that purpose as it would reset the whole
pipeline including the entities list that have been setup at init time.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>