serial: sh-sci: Fix transfer sequence of unsupport DMA transfer
In several R-Car SoC, the channel which is supporting the DMA transfer,
and the channel which is not supporting the DMA transfer are present in
(H)SCIx device.

On this SoC, when a DMA transfer of (H)SCIx device is enabled,
the following errors occur frequently:

    sh-sci xxx: dma_request_slave_channel_compat failed

When the device tree is available, sci_parse_dt () sets the value of
"dmas" property to "dma_slave_tx" and "dma_slave_rx". And
sci_request_dma() does not call dma_request_slave_channel_compat()
regardless of the existence of the device tree, when "dma_slave_tx" or
"dma_slave_rx" is 0.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>