arm64: renesas: r8a7795 dtsi: Update common divider clocks
Update the r8a7795 CPG DTS to include pll1_div2 and pll1_div4.

In the current r8a7795 CPG driver, since the output of PLL1 is divided
by 2 inside dirver from the original value. Therefore, this patch set
div of pll1_div2 as 1.

And, since the parent clock of HDMI and CSI2 was set to pll1_div2,
this patch fixes the parent clock of HDMI and CSI2 to pll1_div4.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>