firmware: psci: Let CPU0 handle Trusted OS by default
Since CPU0 will handle DDR training periodically, it's not allowed
to be turned off via CPU Hotplug processing. While DDR training is
considered as Trusted OS service, the firmware should notify that
information to Linux kernel.

While the firmware is not supported that information yet, let
hard-code that configuration to Linux kernel side.
This is quick hack for current situation. It will be reverted
in future.

Not for upstream merge.

Signed-off-by: Khiem Nguyen <khiem.nguyen.xt@rvc.renesas.com>