pwm: rcar: Add a judgment of the period out of range
Current code does not cover the case that user inputs invalid
value of period.
This patch adds a judgment to output an error when an input period
value is out of the range. As description in hardware manual the
minimum value of period is not smaller than 2 clock cycles.

Signed-off-by: Ryo Kodama <ryo.kodama.vz@renesas.com>
Signed-off-by: Van Do <van.do.xw@renesas.com>
Signed-off-by: Phong Hoang <phong.hoang.wz@renesas.com>
(cherry picked from commit 73d8a0fb46ab4f822f953ad20e5fbdb5998352c1)