PM / Domains: Fix potential deadlock while adding/removing subdomains
We must preserve the same order of how we acquire the lock for
genpd, as otherwise we may encounter deadlocks.

The power on phase of a genpd starts by acquiring its lock. Then it
walks the hierarchy of its parent domains to be able to power on
these first, as per design of genpd.

From the locking perspective this means the locks of the parents
becomes acquired after the lock of the subdomain.

Fix pm_genpd_{add|remove}_subdomain() to maintain the same order
of acquiring the genpd lock as being applied in the power on/off
sequence.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>