[RFC] gpio: rcar: Add Runtime PM handling for interrupts
The R-Car GPIO driver handles Runtime PM for requested GPIOs only.

When using a GPIO purely as an interrupt source, no Runtime PM handling
is done, and the GPIO module's clock may not be enabled.

To fix this:
  - Add .irq_request_resources() and .irq_release_resources() callbacks
    to handle Runtime PM when an interrupt is requested,
  - Add irq_bus_lock() and sync_unlock() callbacks to handle Runtime PM
    when e.g. disabling/enabling an interrupt, or configuring the
    interrupt type.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
---
This fixes ravb Ethernet on r8a7795/Salvator-X, which was "broken" by
commit d5c3d84657db57bd ("net: phy: Avoid polling PHY with
PHY_IGNORE_INTERRUPTS").

Does it also fix the HDMI interrupt on r8a7791/Koelsch?

v2:
  - Handle Runtime PM in .irq_bus_{lock,sync_unlock}() instead of
    .irq_{dis,en}able() and .irq_set_type(),
  - Drop WARN() checks, they were just for testing.
    In case of failures, please re-add

	WARN(pm_runtime_suspended(&p->pdev->dev),
				  "%s: %s is runtime-suspended\n", __func__,
				  dev_name(&p->pdev->dev));

    to gpio_rcar_read() and gpio_rcar_write().