[WIP] gpio: rcar: Always keep the device runtime-resumed
The R-Car GPIO driver handles Runtime PM for requested GPIOs only.

When using a GPIO purely as an interrupt source, no Runtime PM handling
is done, and the GPIO module's clock may not be enabled, leading to
missed interrupts.

Until a proper solution is implemented, fix this by keeping the device
runtime-resumed.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
---
Fixes:
  - RAVB Ethernet on r8a7795/salvator-x, since commit d5c3d84657db57bd
    ("net: phy: Avoid polling PHY with PHY_IGNORE_INTERRUPTS"),
  - HDMI on r8a7791/koelsch (TBC).