thermal: rcar: Add the emergency shutdown control for R-CAR Gen3
Emergency shutdown is the specific function for R-CAR Gen3. In case
of high temperature, it provides cooling system with cpu hotplug.

This function is enabled by default. And it would be activated when
temperature is 110 degrees over. Also the cpu frequency (Cortex-A57
only) is set to minimum frequency.

Please note that performance is declined when it is activated.

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>
Signed-off-by: Hien Dang <hien.dang.eb@rvc.renesas.com>