thermal: rcar_gen3_thermal: Add support for r8a7796
Most of register and operation is the same with r8a7795
except control/status register and initialization flow.
Therefore, new compatible structure is added for r8a7796.
That structure is not only for the initialization flow
but also for future function it might have.

It is assumed that FUSE values are not programmed in
M3 ES10 as H3 WS10/WS11. Consequently, fixed valued
will be used.

It is also assumed that M3 ES10 is supported by
PRR helper function.

Signed-off-by: Hien Dang <hien.dang.eb@rvc.renesas.com>