arm64: dts: r8a7795: Fix the power domain property of VDPB device node
The A2VC1 power domain is set to the power domain property of VDPB
device node. However, the register of VDPB device cannot be read even if
it enables A2VC1 power domain.

This patch fixes the power domain property of the VDPB device node from
A2VC1 to A2VC0, and makes reading of a register possible.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>