thermal: rcar_gen3_thermal: Modify the way to detect the interrupts
The current implementation is that the interrupt I/F and thermal
sensor are assigned with one-to-one. So it can't be set the more
interrupt trigger to one thermal sensor. Also, the interrupt is
detected by a little bit change in temperature.

In order to solve the above problems, the interrupt of thermal
sensor is changed as below.
  - Change the shared interrupt in each thermal sensors.
  - Detect the interrupt when the temperature is changed
    one degree up and down.

Note:
This patch supports 'linear formula calculation'.

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>