clk: renesas: cpg-mssr: Change to combine more than one clocks to device
In current source of clock driver, it intends to bind
one pm clock with a device. So RuntimePM just control
one module stop for one device only.

Driver should implement so that only one module clock
is mapped with one device for aligning with upstream.

In fact, there may be more than one clock are bound with a
device, and driver has not updated to cache up with upstream.

So, this patch is intended for backward adaptability with
current drivers.

Signed-off-by: Dien Pham <dien.pham.ry@rvc.renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>