clk: renesas: rcar-gen3: Use common LSI checking helper functions for WA
The local code for checking LSI version has been added at
early phase of development.
And now, common checking LSI version helper functions are
created for using commonly in system.

It is not good if there are 2 functions exist for same purpose.
So, this patch removes local code to use system common code for
unifying in system.

Signed-off-by: Dien Pham <dien.pham.ry@rvc.renesas.com>