ARM: dts: lager: rename and reindex i2cexio
The rename from i2cexio to i2cexio0 is in in preparation for adding
i2cexio1 which will use the dmuxer for IIC1/I2C1.

The reindexing from i2c8 to i2c10 is to allow space for grouping of
additional GPIO buses to added by follow-up patches to support demuxing of
other i2c buses.

Also note that fallback to GPIO is not provided by the hardware for IIC1/I2C1.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
---
v2:
* Reindex i2c
* Add note that GPIO fallback is not provided by hardware