iommu/ipmmu-vmsa: Reuse iommu groups
Extend the IPMMU driver to group devices together based
on the IPMMU device they are connected to. The slave
devices are kept on a list which is used to determine
if existing groups (and contexts) exist.

With this patch the two SYS-DMAC devices using IPMMU-DS0
on r8a7795 will be grouped together.

Signed-off-by: Magnus Damm <damm+renesas@opensource.se>
(cherry picked from commit 7c36c8eafeb3fb648ea0b053bf0472076f44bf9c)
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>