drm: rcar-du: Fix power management procedure
When executing rcar_du_crtc_suspend() and rcar_du_crtc_resume() in
the power management procedure,
the power management of VSP also would have been executed.
Fix to enable clock of display only in the procedure of resume.
And fix to disable clock of display only in the procedure of suspend.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>