iommu/ipmmu-vmsa: Do backup/restore for all IPMMU context
The processing will do the backup/restore for all IPMMU context in
which are in-use at suspend time.
The processing relies on the in-use context (or IPMMU domain).
Then, perform the backup accordingly.

The backup/restore for UTLB/ASID in IP masters will be done next
separated patch.

NOTE:
 - For IPMMU-VMSA driver, it's in-efficient if the generic backup/
restore functions are applied,
due to the wide range of registers, with discreate register ranges,
esp. for uTLB.
 - Use new ipmmu_ctx_* functions introduced in 'commit d574893aee99
("iommu/ipmmu-vmsa: Write IMCTR twice")'.

Signed-off-by: Khiem Nguyen <khiem.nguyen.xt@rvc.renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Hai Nguyen Pham <hai.pham.ud@renesas.com>