arm64: dts: r8a7796: Update property for Lossy reserved area
The "reusable" property allows kernel to put migratable cached data into
reserved memory in case that the reserved memory has not been used by
the assigned kernel module.

However, the Lossy area could not be applied this property due to its
specialized support for dedicated IP with FCP.
In other words, the integrity of read/write data from CPU can not be
guaranteed which can cause impact on system operation.

Therefore, change the property of Lossy reserved area to "no-map"
to prevent unwanted access from kernel side to that area in all cases.

Signed-off-by: Khiem Nguyen <khiem.nguyen.xt@rvc.renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>