usb: renesas_usbhs: Fix issue when hotplug workqueue run before resume
With below setting and calling flow,
after system resumes from System Suspend To RAM,
the hotplug workqueue will run and
it makes CFIFO register setting before usbhsc_resume() run.
And timeout error will occur in usbhsf_fifo_select().

So, this patch tries to avoid above situation.

Setting:
	INIT_DELAYED_WORK(&priv->notify_hotplug_work, usbhsc_notify_hotplug);
	dfunc->notify_hotplug	= usbhsc_drvcllbck_notify_hotplug;

Calling flow:
	usbhsc_drvcllbck_notify_hotplug()
	usbhsc_notify_hotplug  (go through priv->notify_hotplug_work)
	usbhsc_hotplug()
	usbhs_mod_call(priv, start, priv)
	usbhsg_start()
	usbhsg_try_start()
	usbhs_pipe_init()
	usbhs_pipe_clear()
	usbhs_fifo_clear_dcp()
	usbhsf_fifo_select()

Rework:
If it will prevent the hotplug notification during suspend, it should
be code that makes it easy to understand and general. This rework adds
the suspend status in private structure of usbhs. And it will prevent
the hotplug notification with using it.

Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>
Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>