v4l: vsp1: Use FCP device for DisplayList and VB2 queue
Incrementally fix up parts of the code not yet covered by the
IOMMU patches by Laurent:

[PATCH 0/6] R-Car DU: Fix IOMMU operation when connected to VSP

This patch simply uses the recently introduced function
rcar_fcp_get_device() on the VSP device to retrieve the
FCP device that needs to be used with the DMA MAP API
when IOMMU is enabled.

Without this patch the DU/VSP/FCP devices on R-Car Gen3 will
generate traps during boot when IOMMU is enabled.

Signed-off-by: Magnus Damm <damm+rene...@opensource.se>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>