drm: Add data transmission order bus flag
The flags indicate whether data is transmitted lsb to msb or msb to lsb
on the bus.

The exact meaning is bus-type dependent. For instance, for LVDS buses
the flags indicate whether the seven data bits transmitted in a clock
pulse are sent in normal order (msb to lsb, slots 0 to 6) or reverse
order (lsb to msb, slots 6 to 0).

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
---
Changes since v2:

- Rename the flag to DRM_BUS_FLAG_DATA_LSB_TO_MSB and add a
  corresponding DRM_BUS_FLAG_DATA_MSB_TO_LSB flag.